# remove unused `data-key` attribute