--- a/svelte/package.json
+++ b/svelte/package.json
@@ -10,14 +10,13 @@
     "test": "node --test"
   },
   "dependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.5.2",
-    "fastify": "^4.24.3",
-    "svelte": "^3.59.2"
+    "fastify": "^4.24.3"
   },
   "devDependencies": {
     "@babel/core": "^7.23.3",
     "@babel/eslint-parser": "^7.23.3",
     "@fastify/vite": "^6.0.7",
+    "@sveltejs/vite-plugin-svelte": "4.0.0-next.6",
     "eslint": "^8.54.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.29.0",
@@ -25,6 +24,7 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-svelte": "^2.35.0",
+    "svelte": "5.0.0-next.240",
     "vite": "^5.0.0"
   }
 }