… svelte/pnpm-lock.yaml