I figure if #6 can do it, so can we — another 22% speed boost

Before:

```
┌─────────┬──────┬───────┬───────┬───────┬──────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max   │
├─────────┼──────┼───────┼───────┼───────┼──────────┼─────────┼───────┤
│ Latency │ 9 ms │ 10 ms │ 21 ms │ 22 ms │ 10.92 ms │ 3.37 ms │ 37 ms │
└─────────┴──────┴───────┴───────┴───────┴──────────┴─────────┴───────┘
┌───────────┬────────┬────────┬────────┬────────┬────────┬─────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%  │ Avg    │ Stdev   │ Min    │
├───────────┼────────┼────────┼────────┼────────┼────────┼─────────┼────────┤
│ Req/Sec   │ 773    │ 773    │ 881    │ 901    │ 871.7  │ 36.56   │ 773    │
├───────────┼────────┼────────┼────────┼────────┼────────┼─────────┼────────┤
│ Bytes/Sec │ 214 MB │ 214 MB │ 244 MB │ 250 MB │ 241 MB │ 10.1 MB │ 214 MB │
└───────────┴────────┴────────┴────────┴────────┴────────┴─────────┴────────┘
```

After:

```
┌─────────┬──────┬──────┬───────┬───────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%   │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼───────┼─────────┼─────────┼───────┤
│ Latency │ 7 ms │ 8 ms │ 17 ms │ 17 ms │ 8.92 ms │ 2.84 ms │ 32 ms │
└─────────┴──────┴──────┴───────┴───────┴─────────┴─────────┴───────┘
┌───────────┬────────┬────────┬────────┬────────┬──────────┬─────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%  │ Avg      │ Stdev   │ Min    │
├───────────┼────────┼────────┼────────┼────────┼──────────┼─────────┼────────┤
│ Req/Sec   │ 944    │ 944    │ 1,077  │ 1,088  │ 1,062.19 │ 40.2    │ 944    │
├───────────┼────────┼────────┼────────┼────────┼──────────┼─────────┼────────┤
│ Bytes/Sec │ 226 MB │ 226 MB │ 258 MB │ 261 MB │ 255 MB   │ 9.65 MB │ 226 MB │
└───────────┴────────┴────────┴────────┴────────┴──────────┴─────────┴────────┘
```