@ryansolid 2024-08-28
Yeah Solid keys by reference not by ID. That being said Svelte keys by #each directive argument not by data attribute. And React wouldn't serialize it's key so the whole thing seemed a bit off.